feat(schemas): split generated types into modules
Because:

- we have previously been generating a single `index.d.ts` declaration
  file for *all* schemas types;
- Nimbus Desktop would like to vendor these types, but there are a lot
  of types in this file that will go unused;
- we have duplicate `SetPref1` and `SetPref` types; and
- the logic for writing files was split across multiple CLI options

this commit:

- splits the generated types into multiple modules;
- exports the top-level modules from each at the top-level namespace --
  we can't re-export * from each package due to conflicts;
- rewrites the generated types to replace `SetPref1` with `SetPref`;
- simplifies the CLI options to a single option that is the output
  directory and imposes a file heirarchy; and
- releases these changes as v3005.


Fixes #16934
Fixes #16935
