+++ schemas/VERSION
@@ -1,1 +1,1 @@
-3004.0.0
+3005.0.0
