+++ schemas/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mozilla/nimbus-schemas",
-  "version": "3004.0.0",
+  "version": "3005.0.0",
   "description": "Schemas used by Mozilla Nimbus and related projects.",
   "main": "index.d.ts",
   "repository": {
@@ -19,6 +19,7 @@
   },
   "files": [
     "index.d.ts",
-    "schemas"
+    "schemas",
+    "types"
   ]
 }
